# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Krizový manažerský dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface EmailAccount {
  id: string;
  /** Přátelský název účtu (např. "Práce") */
  label: string;
  email: string;
  imapHost: string;
  imapPort: number;
  createdAt: string;
}

export interface CreateAccountBody {
  /** Přátelský název účtu */
  label: string;
  /** E-mailová adresa */
  email: string;
  /** Heslo k e-mailu */
  password: string;
  /** IMAP server (např. imap.gmail.com) */
  imapHost: string;
  /** IMAP port (výchozí 993) */
  imapPort?: number;
}

export type TaskCategory = (typeof TaskCategory)[keyof typeof TaskCategory];

export const TaskCategory = {
  Finance: "Finance",
  Klient: "Klient",
  Dodavatel: "Dodavatel",
  Operativa: "Operativa",
} as const;

export interface Task {
  id: string;
  accountId: string;
  accountEmail: string;
  subject: string;
  from: string;
  date: string;
  /**
   * @minimum 1
   * @maximum 10
   */
  priority: number;
  category: TaskCategory;
  summary: string;
  draftReply: string;
  createdAt: string;
}

export interface SyncRequest {
  /** ID konkrétního účtu (pokud není zadáno, synchronizují se všechny) */
  accountId?: string;
}

export interface SyncResult {
  /** Počet nově zpracovaných e-mailů */
  synced: number;
  /** Počet synchronizovaných účtů */
  accounts: number;
  errors: string[];
}

export type DashboardStatsByCategory = { [key: string]: number };

export type DashboardStatsByPriority = { [key: string]: number };

export interface DashboardStats {
  total: number;
  /** Počet úkolů s prioritou 8+ */
  burning: number;
  byCategory: DashboardStatsByCategory;
  byPriority: DashboardStatsByPriority;
  accounts: number;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}