# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Krizový manažerský dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: Výpis e-mailových účtů
      description: Vrátí seznam všech nakonfigurovaných e-mailových účtů
      responses:
        "200":
          description: Seznam účtů
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailAccount"

    post:
      operationId: createAccount
      tags: [accounts]
      summary: Přidání nového e-mailového účtu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountBody"
      responses:
        "201":
          description: Účet přidán
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailAccount"
        "400":
          description: Neplatné vstupní údaje
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /accounts/{id}:
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Odstranění e-mailového účtu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Účet odstraněn
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Účet nenalezen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    EmailAccount:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
          description: Přátelský název účtu (např. "Práce")
        email:
          type: string
        imapHost:
          type: string
        imapPort:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - label
        - email
        - imapHost
        - imapPort
        - createdAt
    CreateAccountBody:
      type: object
      properties:
        label:
          type: string
          description: Přátelský název účtu
        email:
          type: string
          description: E-mailová adresa
        password:
          type: string
          description: Heslo k e-mailu
        imapHost:
          type: string
          description: IMAP server (např. imap.gmail.com)
        imapPort:
          type: integer
          description: IMAP port (výchozí 993)
          default: 993
      required:
        - label
        - email
        - password
        - imapHost
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error