# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Krizový manažerský dashboard API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Vrátí seznam všech nakonfigurovaných e-mailových účtů
 * @summary Výpis e-mailových účtů
 */
export const ListAccountsResponseItem = zod.object({
  id: zod.string(),
  label: zod.string().describe('Přátelský název účtu (např. \"Práce\")'),
  email: zod.string(),
  imapHost: zod.string(),
  imapPort: zod.number(),
  createdAt: zod.string(),
});
export const ListAccountsResponse = zod.array(ListAccountsResponseItem);

/**
 * @summary Přidání nového e-mailového účtu
 */
export const createAccountBodyImapPortDefault = 993;

export const CreateAccountBody = zod.object({
  label: zod.string().describe("Přátelský název účtu"),
  email: zod.string().describe("E-mailová adresa"),
  password: zod.string().describe("Heslo k e-mailu"),
  imapHost: zod.string().describe("IMAP server (např. imap.gmail.com)"),
  imapPort: zod
    .number()
    .default(createAccountBodyImapPortDefault)
    .describe("IMAP port (výchozí 993)"),
});

/**
 * @summary Odstranění e-mailového účtu
 */
export const DeleteAccountParams = zod.object({
  id: zod.coerce.string(),
});

export const DeleteAccountResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Výpis úkolů seřazených podle priority
 */
export const listTasksResponsePriorityMax = 10;

export const ListTasksResponseItem = zod.object({
  id: zod.string(),
  accountId: zod.string(),
  accountEmail: zod.string(),
  subject: zod.string(),
  from: zod.string(),
  date: zod.string(),
  priority: zod.number().min(1).max(listTasksResponsePriorityMax),
  category: zod.enum(["Finance", "Klient", "Dodavatel", "Operativa"]),
  summary: zod.string(),
  draftReply: zod.string(),
  createdAt: zod.string(),
});
export const ListTasksResponse = zod.array(ListTasksResponseItem);

/**
 * @summary Označení úkolu jako vyřízený (smazání)
 */
export const DeleteTaskParams = zod.object({
  id: zod.coerce.string(),
});

export const DeleteTaskResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * Stáhne e-maily ze všech nakonfigurovaných účtů a analyzuje je pomocí AI
 * @summary Spuštění synchronizace e-mailů
 */
export const SyncEmailsBody = zod.object({
  accountId: zod
    .string()
    .optional()
    .describe(
      "ID konkrétního účtu (pokud není zadáno, synchronizují se všechny)",
    ),
});

export const SyncEmailsResponse = zod.object({
  synced: zod.number().describe("Počet nově zpracovaných e-mailů"),
  accounts: zod.number().describe("Počet synchronizovaných účtů"),
  errors: zod.array(zod.string()),
});

/**
 * Vrátí přehled počtu úkolů podle priority a kategorie
 * @summary Statistiky dashboardu
 */
export const GetStatsResponse = zod.object({
  total: zod.number(),
  burning: zod.number().describe("Počet úkolů s prioritou 8+"),
  byCategory: zod.record(zod.string(), zod.number()),
  byPriority: zod.record(zod.string(), zod.number()),
  accounts: zod.number(),
});